post_service/requirements.txt
flask
flask_sqlalchemy
pyjwt
flask-cors